requirements.txt
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.6
comm==0.1.2
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
fonttools==4.25.0
frozendict==2.3.0
frozendict==2.3.0
html5lib==1.1
idna==3.4
ipykernel==6.19.2
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.6
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.1.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-server==1.23.4
jupyterlab==3.5.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.22.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==0.8.4
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
multitasking==0.0.9
munkres==1.1.4
nbclassic==0.5.5
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.3
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
platformdirs==2.5.2
ply==3.11
pooch==1.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.36
psutil==5.9.0
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrate-limiter==2.10.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
pywin32==305.1
pywinpty==2.0.10
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.2.0
requests==2.29.0
requests-cache==0.4.13
requests-futures==1.0.0
requests-ratelimiter==0.4.0
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==66.0.0
sip==6.6.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.4
stack-data==0.2.0
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.38.4
widgetsnbextension==4.0.5
win-inet-pton==1.1.0
XlsxWriter==3.1.1
yahooquery==2.3.1
yfinance==0.2.18
